+++ examples/script_generator_graph/ollama/script_multi_generator_ollama.py
@@ -2,8 +2,6 @@
 Basic example of scraping pipeline using ScriptCreatorGraph
 """
 
-import os
-
 from dotenv import load_dotenv
 
 from scrapegraphai.graphs import ScriptCreatorMultiGraph