+++ scrapegraphai/graphs/json_scraper_multi_graph.py
@@ -49,7 +49,6 @@ class JSONScraperMultiGraph(AbstractGraph):
         config: dict,
         schema: Optional[Type[BaseModel]] = None,
     ):
-
         self.copy_config = safe_deepcopy(config)
         self.copy_schema = deepcopy(schema)