+++ scrapegraphai/graphs/smart_scraper_multi_lite_graph.py
@@ -55,7 +55,6 @@ class SmartScraperMultiLiteGraph(AbstractGraph):
         config: dict,
         schema: Optional[Type[BaseModel]] = None,
     ):
-
         self.copy_config = safe_deepcopy(config)
         self.copy_schema = deepcopy(schema)
         super().__init__(prompt, config, source, schema)