+++ scrapegraphai/graphs/xml_scraper_multi_graph.py
@@ -49,7 +49,6 @@ class XMLScraperMultiGraph(AbstractGraph):
         config: dict,
         schema: Optional[Type[BaseModel]] = None,
     ):
-
         self.copy_config = safe_deepcopy(config)
         self.copy_schema = deepcopy(schema)
         super().__init__(prompt, config, source, schema)