+++ scrapegraphai/nodes/parse_node.py
@@ -122,7 +122,7 @@ class ParseNode(BaseNode):
         state.update({self.output[0]: chunks})
         state.update({"parsed_doc": chunks})
         state.update({"content": chunks})
-        
+
         if self.parse_urls:
             state.update({self.output[1]: link_urls})
             state.update({self.output[2]: img_urls})