+++ tests/utils/copy_utils_test.py
@@ -1,5 +1,3 @@
-import copy
-
 import pytest
 from pydantic.v1 import BaseModel